+++ src/ResourceManager/Compute/Commands.Compute/Strategies/Network/NetworkInterfaceStrategy.cs
@@ -37,12 +37,14 @@ namespace Microsoft.Azure.Commands.Compute.Strategies.Network
             string name,
             INestedResourceConfig<Subnet, VirtualNetwork> subnet,
             IResourceConfig<PublicIPAddress> publicIPAddress,
-            IResourceConfig<NetworkSecurityGroup> networkSecurityGroup = null)
+            IResourceConfig<NetworkSecurityGroup> networkSecurityGroup = null,
+            bool enableAcceleratedNetworking = false)
             => Strategy.CreateResourceConfig(
                 resourceGroup: resourceGroup,
                 name: name,
                 createModel: engine => new NetworkInterface
                 {
+                    EnableAcceleratedNetworking = enableAcceleratedNetworking,
                     IpConfigurations = new []
                     {
                         new NetworkInterfaceIPConfiguration