+++ src/ResourceManager/Compute/Commands.Compute/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 ## Current Release
 * `New-AzureRmVm` and `New-AzureRmVmss` are updated to use the latest `Strategies` library.
+* Update SimpleParameterSet of `New-AzureRmVm` to enable Accelerated Network on qualifying vms.
 * Update New-AzureRmVmss simple parameter set to fail creating the vmss when a user specified LB already exists.
 * Update example for New-AzureRmDisk
 * Add example for `New-AzureRmVM`